fix: Metal GPU shader correctness issues
CRITICAL:
- Fix outlier TurboQuant QJL guard: add use_qjl check so QJL sign
  packing is only applied when n_levels < 16 (Algorithm 2). Previously
  unconditionally applied, corrupting K cache codebook indices when
  outlier_bits >= 4 (the default config).
- Add seq_pos + token_count <= max_seq_len bounds check in run_pipeline
  and run_pipeline_calibration to prevent GPU out-of-bounds writes.

HIGH:
- Promote elementwise residual_add to float arithmetic (was fp16).
- Recompute residual in float in fused_residual_norm step 4 instead of
  reading back half-quantized value from global memory.
- Add outlier_bits >= 2 validation in init_outlier to prevent u8
  underflow.
- Fix QJL coefficient comment (was sqrt(2/pi), code is sqrt(pi/2)).
- Remove dead encode_turboquant_{cache_write,attention} helpers and
  their incomplete param structs from ops.rs.

MEDIUM:
- Add head_dim == HEAD_DIM runtime guard in attention.metal.
- Add GQA head mapping guard for non-divisible num_heads/num_kv_heads.
- Fix swapped lo_idx/hi_idx variable names in polarquant matvec.
- Add d_outlier_padded/d_non_padded <= HEAD_DIM guards in TurboQuant
  outlier kernels.
- Add head_dim == HEAD_DIM guard in TurboQuant attention kernel.
- Write zeros for out-of-bounds token IDs in embedding shader.
- Promote int4_dequant to float arithmetic for precision consistency.

+++ crates/ironmill-inference/src/metal/shaders/elementwise.metal
@@ -19,7 +19,7 @@ kernel void residual_add(
     uint tid [[thread_position_in_grid]])
 {
     if (tid >= size) return;
-    output[tid] = a[tid] + b[tid];
+    output[tid] = half(float(a[tid]) + float(b[tid]));
 }
 
 // Buffer copy: output[i] = input[i]
